+++ packages/nextjs-mf/src/plugins/NextFederationPlugin.ts
@@ -126,13 +126,12 @@ export class NextFederationPlugin {
       };
     }
 
+    const allowedPaths = ['pages/', 'app/', 'src/pages/', 'src/app/']
+
     //patch next
     compiler.options.module.rules.push({
       test(req: string) {
-        if (
-          req.includes(path.join(compiler.context, 'pages/')) ||
-          req.includes(path.join(compiler.context, 'app/'))
-        ) {
+        if (allowedPaths.some(p => req.includes(path.join(compiler.context, p)))) {
           return /\.(js|jsx|ts|tsx|md|mdx|mjs)$/i.test(req);
         }
         return false;
@@ -150,10 +149,7 @@ export class NextFederationPlugin {
           test(req: string) {
             if (isServer) {
               // server has no common chunk or entry to hoist into
-              if (
-                req.includes(path.join(compiler.context, 'pages/')) ||
-                req.includes(path.join(compiler.context, 'app/'))
-              ) {
+              if (allowedPaths.some(p => req.includes(path.join(compiler.context, p)))) {
                 return /\.(js|jsx|ts|tsx|md|mdx|mjs)$/i.test(req);
               }
             }
@@ -183,15 +179,12 @@ export class NextFederationPlugin {
     if (this._extraOptions.automaticAsyncBoundary) {
       compiler.options.module.rules.push({
         test: (request: string) => {
-          if (
-            request.includes(path.join(compiler.context, 'pages/')) ||
-            request.includes(path.join(compiler.context, 'app/'))
-          ) {
+          if (allowedPaths.some(p => request.includes(path.join(compiler.context, p)))) {
             return /\.(js|jsx|ts|tsx|md|mdx|mjs)$/i.test(request);
           }
           return false;
         },
-        exclude: [/node_modules/, /_document/, /_middleware/, /pages[\\/]api/],
+        exclude: [/node_modules/, /_document/, /_middleware/,/pages[\\/]middleware/, /pages[\\/]api/],
         resourceQuery: (query) => !query.includes('hasBoundary'),
         loader: path.resolve(__dirname, '../loaders/async-boundary-loader'),
       });